Update README and package.json
Well, I learn things when I publish to npm for the first time. You need
some more fields to your package.json, and having the README read as it
would in a GitHub repo is not the same as it should read in your npm
package. Updates language a bit.